--- pages/SettingsPage.tsx
@@ -1,11 +1,0 @@
-// src/pages/SettingsPage.tsx
-import React from 'react';
-
-const SettingsPage: React.FC = () => (
-  <div>
-    <h2>Settings</h2>
-    <p>Settings content here...</p>
-  </div>
-);
-
-export default SettingsPage;
